add schema trait
